Fix travis CI build
We need to use specific versions of the prerequisites from before
when this was merged back into OpenSC and discontinued on this line.
We'll continue to use this line, because it is the one that is
currently supported by Yocto.

Also get rid of appveyor badge, since we are not using it here